LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.fillRect.transform-expected.txt

2d.fillRect.transform
fillRect is affected by transforms
Actual output:
Expected output:


PASS fillRect is affected by transforms
